+++ assets/javascripts/kitten/components/embed/video/index.js
@@ -1,12 +1,9 @@
 import React, { useRef, useState, useEffect } from 'react'
 import styled from 'styled-components'
-import { pxToRem } from '../../../helpers/utils/typography'
-import COLORS from '../../../constants/colors-config'
 import {
   getReactElementsByType,
   getReactElementsWithoutType,
 } from '../../../helpers/react/get-react-elements'
-import { ScreenConfig } from '../../../constants/screen-config'
 import { PlayerIconNext } from '../../../components/graphics/icons-next/player-icon-next'
 import classNames from 'classnames'