+++ src/esm/index.js
@@ -203,6 +203,7 @@ export { FacebookIconNext } from './components/graphics/icons-next/facebook-icon
 export { FacebookIconCircleNext } from './components/graphics/icons-next/facebook-icon-circle-next';
 export { FlashIconNext } from './components/graphics/icons-next/flash-icon-next';
 export { FilterIconNext } from './components/graphics/icons-next/filter-icon-next';
+export { GarbageIconNext } from './components/graphics/icons-next/garbage-icon-next';
 export { GiftIcon } from './components/graphics/icons-next/gift-icon';
 export { GoogleIconNext } from './components/graphics/icons-next/google-icon-next';
 export { HeartIconNext } from './components/graphics/icons-next/heart-icon-next';
@@ -291,7 +292,8 @@ export { DocumentsStatusBox } from './components/interaction/upload/documents-st
 
 export { Modal } from './components/layer/modal';
 export { Modal as ModalNext } from './components/layer/modal-next';
-export { Overlay } from './components/layer/overlay'; // Layout
+export { Overlay } from './components/layer/overlay';
+export { SlideModal } from './components/layer/slide-modal'; // Layout
 // ------
 
 export { AdaptableGrid, AdaptableGridCol } from './components/layout/adaptable-grid';